Protecting Video Routes

+++ controllers/videoController.js
@@ -61,13 +61,16 @@ export const getEditVideo = async(req, res) => {
     const {params: {id}} = req;
     try{
         const video = await Video.findById(id);
-        res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
-
+        if(video.creator !== req.user.id){
+            throw Error();
+        }else{
+            res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        }
     }catch(error){
         res.redirect(routes.home);
     }
-    
 }
+
 export const postEditVideo = async(req, res) => {
     const {
         params: {id},
@@ -84,9 +87,14 @@ export const postEditVideo = async(req, res) => {
 export const deleteVideo = async(req, res) => {
     const { params: {id}} = req;
     try{
-        await Video.findOneAndRemove({_id:id});
+        const video = await Video.findById({id});
+        if(video.creator !== req.user.id){
+            throw Error();
+        }else{
+            await Video.findOneAndRemove({_id:id});
+        }
     }catch(error){
+        console.log(error);
     }
-
     res.redirect(routes.home); //비디오 삭제가 성공하던 실패하던 home으로 간다.
 }
